+++ analysis/fluxcalib/cmx-fluxcalib.py
@@ -7,297 +7,13 @@ import desispec.io
 from desispec.calibfinder import CalibFinder
 from desispec.fluxcalibration import isStdStar
 
-datadir = os.path.join(os.getenv('DESI_ROOT'), 'spectro', 'data')
-reduxdir = os.path.join(os.getenv('DESI_ROOT'), 'spectro', 'redux', 'daily', 'exposures')
+rawdata_dir = os.path.join(os.getenv('DESI_ROOT'), 'spectro', 'data')
+#reduxdir = os.path.join(os.getenv('DESI_ROOT'), 'spectro', 'redux', 'daily', 'exposures')
+specprod_dir = os.path.join(os.getenv('DESI_ROOT'), 'spectro', 'redux', 'daily')
 nightwatchdir = os.path.join(os.getenv('DESI_ROOT'), 'spectro', 'nightwatch', 'kpno')
 outdir = os.path.join(os.getenv('DESI_ROOT'), 'spectro', 'redux', 'ioannis', 'cmx', 'fluxcalib')
 
-# --------------------------------------------------
-from redrock import zwarning
-class PlotSpec(object):
-    def __init__(self, targets, templates, zscan, zfit, truth=None, archetypes=False):
-        """TODO: document
-        """
-
-        #- Isolate imports of optional dependencies
-        import matplotlib.pyplot as plt
-
-        #- Only keep targets that are in the zfit table
-        keeptargets = list()
-        keepids = set(zfit['targetid'])
-        for t in targets:
-            if t.id in keepids:
-                keeptargets.append(t)
-
-        self.targets = keeptargets
-        self.templates = templates
-        self.archetypes = archetypes
-        self.zscan = zscan
-        self.zfit = zfit
-        self.itarget = 0
-        self.znum = 0
-        self.smooth = 1
-        self.truth = truth
-        self.targetid_to_itarget = {}
-        for i, t in enumerate(self.targets):
-            self.targetid_to_itarget[t.id] = i
-
-        self._fig = plt.figure()
-        self._ax1 = self._fig.add_subplot(211)
-        self._ax2 = self._fig.add_subplot(212)
-
-        self._cid = self._fig.canvas.mpl_connect('key_press_event',
-            self._onkeypress)
-
-        #- Instructions
-        print("---------------------------------------------------------------"
-            "----------")
-        print("Select window then use keyboard shortcuts to navigate:")
-        print("    up/down arrow: previous/next target")
-        print("    left/right arrow: previous/next redshift fit for this"
-            " target")
-        print("    (d)etails")
-        print("---------------------------------------------------------------"
-            "----------")
-
-        #- Disable some default matplotlib key bindings so that we can use keys
-        #- TODO: cache and reset when done
-        plt.rcParams['keymap.forward'] = ''
-        plt.rcParams['keymap.back'] = ''
-
-        plt.ion()
-        self.plot()
-        print('HACK!!')
-        #plt.show( block=True )
-
-    def _onkeypress(self, event):
-        ### print('key', event.key)
-        if event.key == 'right':
-            self.znum = (self.znum + 1) % self.nznum
-            self.plot(keepzoom=True)
-        elif event.key == 'left':
-            self.znum = (self.znum - 1) % self.nznum
-            self.plot(keepzoom=True)
-        elif event.key == 'down':
-            if self.itarget == min(len(self.targets),len(self.zscan))-1:
-                print('At last target')
-            else:
-                self.znum = 0
-                self.itarget += 1
-                self.plot()
-        elif event.key == 'up':
-            if self.itarget == 0:
-                print('Already at first target')
-            else:
-                self.znum = 0
-                self.itarget -= 1
-                self.plot()
-        elif event.key == 'd':
-            target = self.targets[self.itarget]
-            zfit = self.zfit[self.zfit['targetid'] == target.id]
-            print('target {}'.format(target.id))
-            print(zfit['znum', 'spectype', 'z', 'zerr', 'zwarn', 'chi2'])
-
-    def plot(self, keepzoom=False):
-
-        #- Isolate imports of optional dependencies
-        from scipy.signal import medfilt
-
-        target = self.targets[self.itarget]
-        zfit = self.zfit[self.zfit['targetid'] == target.id]
-        self.nznum = len(zfit)
-        zz = zfit[zfit['znum'] == self.znum][0]
-        coeff = zz['coeff']
-
-        fulltype = zz['spectype']
-        if zz['subtype'] != '':
-            fulltype = fulltype+':::'+zz['subtype']
-        if self.archetypes:
-            dwave = { s.wavehash:s.wave for s in target.spectra }
-            tp = self.archetypes.archetypes[zz['spectype']]
-        else:
-            tp = self.templates[fulltype]
-            if tp.template_type != zz['spectype']:
-                raise ValueError('spectype {} not in'
-                    ' templates'.format(zz['spectype']))
-
-        #----- zscan plot
-        if keepzoom:
-            force_xlim = self._ax1.get_xlim()
-            force_ylim = self._ax1.get_ylim()
-
-        self._ax1.clear()
-        for spectype, fmt in [('STAR', 'k-'), ('GALAXY', 'b-'), ('QSO', 'g-')]:
-            if spectype in self.zscan[target.id]:
-                zx = self.zscan[target.id][spectype]
-                self._ax1.plot(zx['redshifts'], zx['zchi2'], fmt, alpha=0.2,
-                    label='_none_')
-                self._ax1.plot(zx['redshifts'], zx['zchi2']+zx['penalty'], fmt,
-                    label=spectype)
-
-        self._ax1.plot(zfit['z'], zfit['chi2'], 'r.', label='_none_')
-        for row in zfit:
-            self._ax1.text(row['z'], row['chi2'], str(row['znum']),
-                verticalalignment='top')
-
-        if self.truth is not None:
-            i = np.where(self.truth['targetid'] == target.id)[0]
-            if len(i) > 0:
-                ztrue = self.truth['ztrue'][i[0]]
-                self._ax1.axvline(ztrue, color='g', alpha=0.5)
-            else:
-                print('WARNING: target id {} not in truth'
-                    ' table'.format(target.id))
-
-        self._ax1.axvline(zz['z'], color='k', alpha=0.1)
-        self._ax1.axhline(zz['chi2'], color='k', alpha=0.1)
-        self._ax1.legend()
-        self._ax1.set_title('target {}  zbest={:.3f} {}'.format(target.id,
-            zz['z'], zz['spectype']))
-        self._ax1.set_ylabel(r'$\chi^2$')
-        self._ax1.set_xlabel('redshift')
-        if keepzoom:
-            self._ax1.set_xlim(*force_xlim)
-            self._ax1.set_ylim(*force_ylim)
-
-        #----- spectrum plot
-        if keepzoom:
-            force_xlim = self._ax2.get_xlim()
-            force_ylim = self._ax2.get_ylim()
-
-        self._ax2.clear()
-        ymin = ymax = 0.0
-        specs_to_read = target.spectra
-        for spec in specs_to_read:
-            if self.archetypes:
-                mx = tp.eval(zz['subtype'], dwave, coeff, spec.wave, zz['z']) * (1+zz['z'])
-            else:
-                mx = tp.eval(coeff[0:tp.nbasis], spec.wave, zz['z']) * (1+zz['z'])
-            model = spec.R.dot(mx)
-            flux = spec.flux.copy()
-            isbad = (spec.ivar == 0)
-            ## model[isbad] = mx[isbad]
-            flux[isbad] = np.NaN
-            self._ax2.plot(spec.wave, medfilt(flux, self.smooth), alpha=0.5)
-            self._ax2.plot(spec.wave, medfilt(mx, self.smooth), 'k:', alpha=0.8)
-            model[isbad] = np.NaN
-            self._ax2.plot(spec.wave, medfilt(model, self.smooth), 'k-',
-                alpha=0.8)
-
-            if flux[~isbad].size!=0:
-                ymin = min(ymin, np.percentile(flux[~isbad], 1))
-                ymax = max(ymax, np.percentile(flux[~isbad], 99),
-                    np.max(model)*1.05)
-
-        if (ymin==0.) & (ymax==0.):
-             ymax = 1.
-
-        #- Label object type and redshift
-        label = 'znum {} {} z={:.3f}'.format(self.znum, fulltype, zz['z'])
-        print('target {} id {} {}'.format(self.itarget, target.id, label))
-        ytext = ymin+0.9*(ymax-ymin)
-        self._ax2.text(3800, ytext, label)
-
-        #- ZWARN labels
-        if zz['zwarn'] != 0:
-            label = list()
-            for name, mask in zwarning.ZWarningMask.flags():
-                if (zz['zwarn'] & mask) != 0:
-                    label.append(name)
-            label = '\n'.join(label)
-            color = 'r'
-        else:
-            label = 'ZWARN=0'
-            color = 'g'
-
-        self._ax2.text(10000, ytext, label, horizontalalignment='right',
-            color=color)
-
-        self._ax2.axhline(0, color='k', alpha=0.2)
-        if keepzoom:
-            self._ax2.set_xlim(*force_xlim)
-            self._ax2.set_ylim(*force_ylim)
-        else:
-            self._ax2.set_ylim(ymin, ymax)
-            self._ax2.set_xlim(3500,10100)
-
-        self._ax2.set_ylabel('flux')
-        self._ax2.set_xlabel('wavelength [A]')
-        # self._fig.tight_layout()
-        self._fig.canvas.draw()
-# --------------------------------------------------
-
-def gather_qa(night, verbose=False, overwrite=False):
-    """Read and stack all the nightwatch QA files for a given night.
-
-    """
-    import json
-    import astropy.table
-    from astropy.table import Table
-
-    stackwatchfile = os.path.join(outdir, 'qa-nightwatch-{}.fits'.format(str(night)))
-    fiberassignmapfile = os.path.join(outdir, 'fiberassignmap-{}.fits'.format(str(night)))
-
-    if os.path.isfile(stackwatchfile) and not overwrite:
-        print('Reading {}'.format(stackwatchfile))
-        data = Table(fitsio.read(stackwatchfile, 'PER_CAMFIBER'))
-        fiberassignmap = Table(fitsio.read(fiberassignmapfile))
-    else:
-        #print('Reading the focal plane model.')
-        #fp = desimodel.io.load_focalplane()[0]
-        #fp = fp['PETAL', 'FIBER', 'OFFSET_X', 'OFFSET_Y']
-
-        nightdir = os.path.join(nightwatchdir, str(night))
-        allexpiddir = glob(os.path.join(nightdir, '????????'))
-
-        data = []
-        fiberassignmap = Table(names=('NIGHT', 'EXPID', 'TILEID', 'FIBERASSIGNFILE'),
-                               dtype=('U8', 'U8', 'i4', 'U32'))
-        for expiddir in allexpiddir:
-            expid = os.path.basename(expiddir)
-            qafile = os.path.join(expiddir, 'qa-{}.fits'.format(expid))
-
-            qaFITS = fitsio.FITS(qafile)
-            if 'PER_CAMFIBER' in qaFITS:
-                if verbose:
-                    print('Reading {}'.format(qafile))
-                qa = Table(qaFITS['PER_CAMFIBER'].read())
-
-            # Hack! Figure out the mapping between EXPID and FIBERMAPusing the request-EXPID.json file.
-            requestfile = os.path.join(datadir, str(night), expid, 'request-{}.json'.format(expid))
-            with open(requestfile) as ff:
-                req = json.load(ff)
-            if 'PASSTHRU' in req.keys():
-                tileid = int(req['PASSTHRU'].split(':')[3].split(',')[0])
-                tilefile = glob(os.path.join(datadir, str(night), '????????', 'fiberassign-{:06d}.fits'.format(tileid)))
-                #if len(tilefile) == 0:
-                #    print('No fibermap file found for EXPID={}'.format(expid))
-                #if len(tilefile) > 0:
-                #    print('Multiple fibermap files found for EXPID={}!'.format(expid))
-                if len(tilefile) > 0:
-                    tsplit = tilefile[0].split('/')
-                    fiberassignmap.add_row((str(night), str(expid), tileid, os.path.join(tsplit[-2], tsplit[-1])))
-                    #fiberassignmap[str(expid)] = [tileid]
-                    data.append(qa)
-        data = astropy.table.vstack(data)
-
-        # Need to update the data model to 'f4'.
-        print('Updating the data model.')
-        for col in data.colnames:
-            if data[col].dtype == '>f8':
-                data[col] = data[col].astype('f4')
-
-        print('Writing {}'.format(stackwatchfile))
-        fitsio.write(stackwatchfile, data.as_array(), clobber=True, extname='PER_CAMFIBER')
-
-        print('Writing {}'.format(fiberassignmapfile))
-        # ValueError: unsupported type 'U42'
-        #fitsio.write(fiberassignmapfile, fiberassignmap.as_array(), clobber=True)
-        fiberassignmap.write(fiberassignmapfile, overwrite=True)
-
-    return data, fiberassignmap
-
-def select_stdstars(data, fiberassignmap, snrcut=10, verbose=False):
+def select_stdstars(data, fiberassignmap, snrcut=5, verbose=False):
     """Select spectra based on S/N and targeting bit.
 
     """
@@ -308,7 +24,7 @@ def select_stdstars(data, fiberassignmap, snrcut=10, verbose=False):
 
     tileid = fiberassignmap['TILEID'][0]
     fibermapfile = fiberassignmap['FIBERASSIGNFILE'][0]
-    fibermapfile = os.path.join(datadir, str(night), fibermapfile)
+    fibermapfile = os.path.join(rawdata_dir, str(night), fibermapfile)
     if verbose:
         print('Reading {}'.format(fibermapfile))
     fibermap = Table(fitsio.read(fibermapfile))#, columns=['CMX_TARGET']))
@@ -349,14 +65,102 @@ def select_stdstars(data, fiberassignmap, snrcut=10, verbose=False):
 
     return fibermap
 
-def update_frame_fibermaps(data, fiberassignmap, verbose=False, overwrite=False):
+def gather_qa(night, verbose=False, overwrite=False):
+    """Read and stack all the nightwatch QA files for a given night.
+
+    """
+    import json
+    import astropy.table
+    from astropy.table import Table
+
+    stackwatchfile = os.path.join(outdir, 'qa-nightwatch-{}.fits'.format(str(night)))
+    fiberassignmapfile = os.path.join(outdir, 'fiberassignmap-{}.fits'.format(str(night)))
+
+    if os.path.isfile(stackwatchfile) and not overwrite:
+        print('Reading {}'.format(stackwatchfile))
+        data = Table(fitsio.read(stackwatchfile, 'PER_CAMFIBER'))
+        fiberassignmap = Table(fitsio.read(fiberassignmapfile))
+    else:
+        #print('Reading the focal plane model.')
+        #fp = desimodel.io.load_focalplane()[0]
+        #fp = fp['PETAL', 'FIBER', 'OFFSET_X', 'OFFSET_Y']
+
+        nightdir = os.path.join(nightwatchdir, str(night))
+        allexpiddir = glob(os.path.join(nightdir, '????????'))
+
+        data = []
+        fiberassignmap = Table(names=('NIGHT', 'EXPID', 'TILEID', 'FIBERASSIGNFILE'),
+                               dtype=('U8', 'U8', 'i4', 'U32'))
+        for expiddir in allexpiddir:
+            expid = os.path.basename(expiddir)
+            qafile = os.path.join(expiddir, 'qa-{}.fits'.format(expid))
+
+            qaFITS = fitsio.FITS(qafile)
+            if 'PER_CAMFIBER' in qaFITS:
+                if verbose:
+                    print('Reading {}'.format(qafile))
+                qa = Table(qaFITS['PER_CAMFIBER'].read())
+
+            # Hack! Figure out the mapping between EXPID and FIBERMAPusing the request-EXPID.json file.
+            requestfile = os.path.join(rawdata_dir, str(night), expid, 'request-{}.json'.format(expid))
+            #print(requestfile)
+            with open(requestfile) as ff:
+                req = json.load(ff)
+            if 'PASSTHRU' in req.keys():
+                if type(req['PASSTHRU']) is dict:
+                    tileid = req['PASSTHRU']['TILEID']
+                    #tileid = int(req['PASSTHRU'].split(':')[3].split(',')[0])
+                else:
+                    indx = req['PASSTHRU'].index('TILEID')
+                    tileid = req['PASSTHRU'][indx:]
+                    tileid = int(tileid[tileid.index(':')+1:tileid.index(',')])
+                tilefile = glob(os.path.join(rawdata_dir, str(night), '????????', 'fiberassign-{:06d}.fits'.format(tileid)))
+                #if len(tilefile) == 0:
+                #    print('No fibermap file found for EXPID={}'.format(expid))
+                #if len(tilefile) > 0:
+                #    print('Multiple fibermap files found for EXPID={}!'.format(expid))
+                if len(tilefile) > 0:
+                    tsplit = tilefile[0].split('/')
+                    fiberassignmap.add_row((str(night), str(expid), tileid, os.path.join(tsplit[-2], tsplit[-1])))
+                    #fiberassignmap[str(expid)] = [tileid]
+                    data.append(qa)
+                #else:
+                #    print('  No tilefile found')
+            #else:
+            #    print('  No tilefile found')
+
+        if len(data) == 0:
+            print('No fiberassign files found for night {}'.format(night))
+            return None, None
+        data = astropy.table.vstack(data)
+
+        # Need to update the data model to 'f4'.
+        print('Updating the data model.')
+        for col in data.colnames:
+            if data[col].dtype == '>f8':
+                data[col] = data[col].astype('f4')
+
+        print('Writing {}'.format(stackwatchfile))
+        fitsio.write(stackwatchfile, data.as_array(), clobber=True, extname='PER_CAMFIBER')
+
+        print('Writing {}'.format(fiberassignmapfile))
+        # ValueError: unsupported type 'U42'
+        #fitsio.write(fiberassignmapfile, fiberassignmap.as_array(), clobber=True)
+        fiberassignmap.write(fiberassignmapfile, overwrite=True)
+
+    return data, fiberassignmap
+
+def update_frame_fibermaps(night, verbose=False, overwrite=False):
     """Update the fibermaps in each frame file with the necessary information on
     targeted standard stars.
 
     """
     from desitarget.targets import main_cmx_or_sv
     
-    night = data['NIGHT'][0]
+    # Build the stacked QA and fiberassign "map" files.
+    data, fiberassignmap = gather_qa(night, verbose=verbose, overwrite=overwrite)
+    if data is None:
+        return
     
     # For each EXPID, find the fibers with standard stars *and* good spectra.
     for expid in set(data['EXPID']):
@@ -369,9 +173,11 @@ def update_frame_fibermaps(data, fiberassignmap, verbose=False, overwrite=False)
         #print(np.sum(fibermap['CMX_TARGET'] & cmx_mask.mask('STD_BRIGHT') != 0))
 
         strexpid = '{:08d}'.format(expid)
-        framefiles = glob(os.path.join(reduxdir, str(night), strexpid, 'frame-[brz][0-9]-{}.fits'.format(strexpid)))
+        framefiles = sorted(glob(desispec.io.findfile(
+            'frame', night=night, expid=expid,
+            specprod_dir=specprod_dir, camera='*')))
         for framefile in framefiles:
-            outframefile = os.path.join(outdir, 'exposures', str(night), strexpid, os.path.basename(framefile))
+            outframefile = framefile.replace(specprod_dir, outdir)
             if os.path.isfile(outframefile) and not overwrite:
                 print('File exists {}'.format(outframefile))
                 continue
@@ -399,81 +205,96 @@ def update_frame_fibermaps(data, fiberassignmap, verbose=False, overwrite=False)
                     print('Writing {}'.format(outframefile))
                 desispec.io.write_frame(outframefile, fr)
 
-        print('Returning after first EXPID - fix me.')
-        return
+        if False:
+            print('Returning after first EXPID - fix me.')
+            return
 
-def fit_stdstars(night, verbose=False, overwrite=False):
-    """Fit the standards on each petal.
+def fluxcalib(night, verbose=False, overwrite=False):
+    """Fit the standards on each petal and then perform flux-calibration.
 
     desi_fit_stdstars --frames 00028833/frame-*.fits --skymodels 00028833/sky-*.fits \
       --fiberflats $DESI_SPECTRO_CALIB/spec/sp3/fiberflat-sm4-*-20191108.fits \
       --starmodels $DESI_BASIS_TEMPLATES/stdstar_templates_v2.2.fits --outfile stdstars-3-00028833.fits
-
-    """
-    starmodelfile = os.path.join(os.getenv('DESI_BASIS_TEMPLATES'), 'stdstar_templates_v2.2.fits')
-
-    allexpiddir = glob(os.path.join(outdir, 'exposures', str(night), '????????'))
-    for expiddir in allexpiddir:
-        expid = os.path.basename(expiddir)
-        # Process each spectrograph separately.
-        for spectro in np.arange(10):
-            outstdfile = os.path.join(expiddir, 'stdstars-{}-{}.fits'.format(str(spectro), expid))
-            if os.path.isfile(outstdfile) and not overwrite:
-                print('File exists {}'.format(outstdfile))
-                continue
-
-            framefiles = sorted(glob(os.path.join(expiddir, 'frame-[brz]{}-{}.fits'.format(str(spectro), expid))))
-            # Gather the calibration files.
-            if len(framefiles) == 3:
-                skymodelfiles = [framefile.replace(outdir, reduxdir).replace('frame-', 'sky-') for framefile in framefiles]
-                fiberflatfiles = []
-                for framefile in framefiles:
-                    calib = CalibFinder([fitsio.read_header(framefile)])
-                    fiberflatfiles.append(os.path.join(os.getenv('DESI_SPECTRO_CALIB'), calib.data['FIBERFLAT']))
-
-                cmd = 'desi_fit_stdstars --frames {framefiles} --skymodels {skymodelfiles} --fiberflats {fiberflatfiles} '
-                cmd += '--starmodels {starmodelfile} --outfile {outstdfile}'
-                cmd = cmd.format(framefiles=' '.join(framefiles),
-                                 skymodelfiles=' '.join(skymodelfiles),
-                                 fiberflatfiles=' '.join(fiberflatfiles),
-                                 starmodelfile=starmodelfile,
-                                 outstdfile=outstdfile)
-                #print(cmd)
-                os.system(cmd)
-
-def fluxcalib(night, verbose=False, overwrite=False):
-    """Perform flux-calibration.
 
     desi_compute_fluxcalibration  --infile 00028833/frame-b3-*.fits --sky 00028833/sky-b3-*.fits \
       --fiberflat $DESI_SPECTRO_CALIB/spec/sp3/fiberflat-sm4-b-20191108.fits --models stdstars-3-00028833.fits \
       --outfile fluxcalib-b3-00028833.fits --delta-color-cut 12
 
     """
+    starmodelfile = os.path.join(os.getenv('DESI_BASIS_TEMPLATES'), 'stdstar_templates_v2.2.fits')
+    
     allexpiddir = glob(os.path.join(outdir, 'exposures', str(night), '????????'))
     for expiddir in allexpiddir:
         expid = os.path.basename(expiddir)
         
         # Process each spectrograph separately.
         for spectro in np.arange(10):
-            modelsfile = os.path.join(expiddir, 'stdstars-{}-{}.fits'.format(str(spectro), expid))
-            if os.path.isfile(modelsfile):
-                framefiles = sorted(glob(os.path.join(expiddir, 'frame-[brz]{}-{}.fits'.format(str(spectro), expid))))
+            framefiles = sorted(glob(desispec.io.findfile(
+                'frame', night=night, expid=int(expid), spectrograph=spectro,
+                camera='*{}'.format(spectro), specprod_dir=outdir)))
+            if len(framefiles) == 0:
+                print('No frame files found for spectrograph {}'.format(spectro))
+                continue
+
+            # Gather the calibration files.
+            havefiles = True if len(framefiles) == 3 else False
+            if havefiles:
+                fiberflatfiles, skymodelfiles = [], []
                 for framefile in framefiles:
-                    outcalibfile = framefile.replace('frame-', 'fluxcalib-')
+                    hdr = fitsio.read_header(framefile)
+                    calib = CalibFinder([hdr])
+                    
+                    fiberflatfile = os.path.join(os.getenv('DESI_SPECTRO_CALIB'), calib.data['FIBERFLAT'])
+                    skymodelfile = desispec.io.findfile(
+                        'sky', night=night, expid=int(expid), spectrograph=spectro,
+                        camera=hdr['CAMERA'].strip(), specprod_dir=specprod_dir) # note: specprod_dir!
+                    if not os.path.isfile(fiberflatfile) or not os.path.isfile(skymodelfile):
+                        havefiles = False
+                    else:
+                        fiberflatfiles.append(fiberflatfile)
+                        skymodelfiles.append(skymodelfile) 
+
+            # Fit the standard stars (per-spectrograph, all three cameras simultaneously).
+            stdstarsfile = desispec.io.findfile('stdstars', night=night, expid=int(expid),
+                                                spectrograph=spectro, specprod_dir=outdir)
+            if os.path.isfile(stdstarsfile) and not overwrite:
+                print('File exists {}'.format(stdstarsfile))
+            else:
+                if havefiles:
+                    cmd = 'desi_fit_stdstars --frames {framefiles} --skymodels {skymodelfiles} '
+                    cmd += '--fiberflats {fiberflatfiles} --starmodels {starmodelfile} '
+                    cmd += '--outfile {stdstarsfile}'
+                    cmd = cmd.format(framefiles=' '.join(framefiles),
+                                     skymodelfiles=' '.join(skymodelfiles),
+                                     fiberflatfiles=' '.join(fiberflatfiles),
+                                     starmodelfile=starmodelfile,
+                                     stdstarsfile=stdstarsfile)
+                    #print(cmd)
+                    os.system(cmd)
+
+            # Perform flux-calibration (per-spectrograph).
+            if os.path.isfile(stdstarsfile) and havefiles:
+                for iframe in np.arange(len(framefiles)):
+                    framefile = framefiles[iframe]
+                    camera = fitsio.read_header(framefile)['CAMERA'].strip()
+
+                    outcalibfile = desispec.io.findfile(
+                        'fluxcalib', night=night, expid=int(expid), spectrograph=spectro,
+                        camera=camera, specprod_dir=outdir)
                     if os.path.isfile(outcalibfile) and not overwrite:
                         print('File exists {}'.format(outcalibfile))
                         continue
                     
-                    # Gather the calibration files.
-                    skyfile = framefile.replace(outdir, reduxdir).replace('frame-', 'sky-')
-                    calib = CalibFinder([fitsio.read_header(framefile)])
-                    fiberflatfile = os.path.join(os.getenv('DESI_SPECTRO_CALIB'), calib.data['FIBERFLAT'])
+                    ## Gather the calibration files.
+                    #skyfile = framefile.replace(outdir, reduxdir).replace('frame-', 'sky-')
+                    #calib = CalibFinder([fitsio.read_header(framefile)])
+                    #fiberflatfile = os.path.join(os.getenv('DESI_SPECTRO_CALIB'), calib.data['FIBERFLAT'])
 
                     cmd = 'desi_compute_fluxcalibration --infile {framefile} --sky {skyfile} '
-                    cmd += '--fiberflat {fiberflatfile} --models {modelsfile} --outfile {outcalibfile} '
+                    cmd += '--fiberflat {fiberflatfile} --models {stdstarsfile} --outfile {outcalibfile} '
                     cmd += '--delta-color-cut 12'
-                    cmd = cmd.format(framefile=framefile, skyfile=skyfile,
-                                     fiberflatfile=fiberflatfile, modelsfile=modelsfile,
+                    cmd = cmd.format(framefile=framefiles[iframe], skyfile=skymodelfiles[iframe],
+                                     fiberflatfile=fiberflatfiles[iframe], stdstarsfile=stdstarsfile,
                                      outcalibfile=outcalibfile)
                     #print(cmd)
                     os.system(cmd)
@@ -507,7 +328,9 @@ def process_exposures(night, verbose=False, overwrite=False):
     for expiddir in allexpiddir:
         expid = os.path.basename(expiddir)
 
-        framefiles = sorted(glob(os.path.join(expiddir, 'frame-[brz]?-{}.fits'.format(expid))))
+        framefiles = sorted(glob(desispec.io.findfile(
+            'frame', night=night, expid=int(expid),
+            camera='*', specprod_dir=outdir)))
         for framefile in framefiles:
             outframefile = framefile.replace('frame-', 'cframe-')
             if os.path.isfile(outframefile) and not overwrite:
@@ -515,20 +338,80 @@ def process_exposures(night, verbose=False, overwrite=False):
                 continue
 
             # Gather the calibration files.
-            skyfile = framefile.replace(outdir, reduxdir).replace('frame-', 'sky-')
-            calib = CalibFinder([fitsio.read_header(framefile)])
+            hdr = fitsio.read_header(framefile)
+            camera = hdr['CAMERA'].strip()
+            calib = CalibFinder([hdr])
+                    
             fiberflatfile = os.path.join(os.getenv('DESI_SPECTRO_CALIB'), calib.data['FIBERFLAT'])
-            calibfile = framefile.replace('frame-', 'fluxcalib-')
+            skymodelfile = desispec.io.findfile('sky', night=night, expid=int(expid),
+                                                camera=camera, specprod_dir=specprod_dir) # note: specprod_dir!
+            fluxcalibfile = desispec.io.findfile('fluxcalib', night=night, expid=int(expid), 
+                                                 camera=camera, specprod_dir=outdir) # note: outdir!
+            if os.path.isfile(fiberflatfile) and os.path.isfile(skymodelfile) and os.path.isfile(fluxcalibfile):
+                cmd = 'desi_process_exposure --infile {framefile} --sky {skymodelfile} '
+                cmd += '--fiberflat {fiberflatfile} --calib {fluxcalibfile} '
+                cmd += '--cosmics-nsig 6 --outfile {outframefile} '
+                cmd = cmd.format(framefile=framefile, skymodelfile=skymodelfile,
+                                 fiberflatfile=fiberflatfile, fluxcalibfile=fluxcalibfile,
+                                 outframefile=outframefile)
+                #print(cmd)
+                os.system(cmd)
+            else:
+                print('Missing calibration files.')
 
-            cmd = 'desi_process_exposure --infile {framefile} --sky {skyfile} '
-            cmd += '--fiberflat {fiberflatfile} --calib {calibfile} '
-            cmd += '--cosmics-nsig 6 --outfile {outframefile} '
-            cmd = cmd.format(framefile=framefile, skyfile=skyfile,
-                             fiberflatfile=fiberflatfile, calibfile=calibfile,
-                             outframefile=outframefile)
-            #print(cmd)
-            os.system(cmd)
+def qa_stdstars(night, verbose=False, overwrite=False):
+    """Make some QAplots of the standard stars.
 
+    """
+    #import matplotlib.pyplot as plt
+    from astropy.table import Table, vstack
+    from desispec.io.fluxcalibration import read_flux_calibration, read_stdstar_models
+    
+    pngfile = os.path.join(outdir, 'qa-stdstars-{}.png'.format(str(night)))
+    if os.path.isfile(pngfile) and not overwrite:
+        print('File exists {}'.format(pngfile))
+        return
+
+    #fig, ax = plt.subplots()
+    
+    allexpiddir = glob(os.path.join(outdir, 'exposures', str(night), '????????'))
+
+    stdmeta = []
+    for expiddir in allexpiddir:
+        expid = os.path.basename(expiddir)
+
+        cframefiles = sorted(glob(desispec.io.findfile(
+            'cframe', night=night, expid=int(expid),
+            camera='*', specprod_dir=outdir)))
+        if len(cframefiles) != 3:
+            continue
+        for cframefile in cframefiles:
+            framefile = cframefile.replace('cframe-', 'frame-')
+
+            from desispec.qa.qa_frame import qaframe_from_frame
+            qaframe_from_frame(framefile, output_dir=outdir, specprod_dir=outdir, make_plots=True, clobber=True)
+            pdb.set_trace()
+            
+            hdr = fitsio.read_header(cframefile)
+            camera = hdr['CAMERA'].strip()
+            if camera[0] == 'b':
+                spectro = int(camera[1])
+                stdstarsfile = desispec.io.findfile('stdstars', night=night, expid=int(expid),
+                                                    spectrograph=spectro, specprod_dir=outdir)
+                print(stdstarsfile)
+                stdflux, stdwave, stdfibers, _stdmeta = read_stdstar_models(stdstarsfile)
+                stdmeta.append(Table(_stdmeta))
+
+            fluxcalibfile = desispec.io.findfile('fluxcalib', night=night, expid=int(expid), 
+                                                 camera=camera, specprod_dir=outdir) # note: outdir!
+            fluxcalib = read_flux_calibration(fluxcalibfile)
+
+            fr = desispec.io.read_frame(cframefile)
+            
+    stdmeta = vstack(stdmeta)
+    pdb.set_trace()
+
+                
 def qa_throughput(night, verbose=False, overwrite=False):
     """Make some QAplots of the throughput.
 
@@ -563,9 +446,10 @@ def qa_throughput(night, verbose=False, overwrite=False):
         expid = os.path.basename(expiddir)
 
         for spectro in np.arange(10):
-            calibfile = glob(os.path.join(expiddir, 'fluxcalib-r{}-{}.fits'.format(spectro, expid)))
-            if len(calibfile) > 0:
-                col = color=next(colors)
+            calibfiles = glob(os.path.join(expiddir, 'fluxcalib-*{}-{}.fits'.format(spectro, expid)))
+            if len(calibfiles) == 3:
+                #print(calibfile, os.path.isfile(calibfile))
+                col = next(colors)
                 for camera in ('b', 'r' , 'z'):
                     cframefile = os.path.join(expiddir, 'cframe-{}{}-{}.fits'.format(camera, spectro, expid))
                     calibfile = os.path.join(expiddir, 'fluxcalib-{}{}-{}.fits'.format(camera, spectro, expid))
@@ -684,12 +568,6 @@ def qa_redrock(night, verbose=False, overwrite=False):
 
         pdb.set_trace()
     
-    
-
-        
-        
-    
-    
     #rrplot --specfile spectra-64/spectra-64-756.fits --rrfile spectra-64/redrock-64-756.h5 --targetids 35186382380993054,611647134688609410
 
     pdb.set_trace()
@@ -698,44 +576,38 @@ def main():
 
     parser = argparse.ArgumentParser(description='Derive flux-calibrated spectra and estimate the throughput.')
 
-    parser.add_argument('-n', '--night', type=int, default=20200102, required=True, help='night')
-    parser.add_argument('--gather-qa', action='store_true', help='Gather and stack nightwatch QA files.')
+    parser.add_argument('-n', '--night', type=int, default=20200102, nargs='*', required=True, help='night')
     parser.add_argument('--update-fibermaps', action='store_true', help='Update fibermap files')
-    parser.add_argument('--fit-stdstars', action='store_true', help='Fit the standard stars.')
-    parser.add_argument('--fluxcalib', action='store_true', help='Do the flux-calibration.')
+    parser.add_argument('--fluxcalib', action='store_true', help='Perform flux-calibration.')
     parser.add_argument('--process-exposures', action='store_true', help='Process all exposures.')
     parser.add_argument('--group-spectra', action='store_true', help='Group the data into healpixels.')
     parser.add_argument('--redrock', action='store_true', help='Do redshift fitting.')
 
     parser.add_argument('--qaplots', action='store_true', help='Make some plots.')
-    parser.add_argument('--verbose', action='store_true', help='Be verbose.')
-    parser.add_argument('--overwrite', action='store_true', help='Overwrite existing files.')
+    parser.add_argument('-o', '--overwrite', action='store_true', help='Overwrite existing files.')
+    parser.add_argument('-v', '--verbose', action='store_true', help='Be verbose.')
     args = parser.parse_args()
-
-    # Gather QA files
-    if args.gather_qa:
-        data, fiberassignmap = gather_qa(args.night, verbose=args.verbose, overwrite=args.overwrite)
 
     # Update fibermap and frame files.
     if args.update_fibermaps:
-        data, fiberassignmap = gather_qa(args.night, verbose=args.verbose)
-        update_frame_fibermaps(data, fiberassignmap, verbose=args.verbose, overwrite=args.overwrite)
+        for night in np.atleast_1d(args.night):
+            update_frame_fibermaps(night, verbose=args.verbose, overwrite=args.overwrite)
 
-    # Fit the standard stars.
-    if args.fit_stdstars:
-        fit_stdstars(args.night, verbose=args.verbose, overwrite=args.overwrite)
-
+    # Fit the standard stars and perform flux-calibration.
     if args.fluxcalib:
-        fluxcalib(args.night, verbose=args.verbose, overwrite=args.overwrite)
+        for night in np.atleast_1d(args.night):
+            fluxcalib(night, verbose=args.verbose, overwrite=args.overwrite)
 
     if args.process_exposures:
-        process_exposures(args.night, verbose=args.verbose, overwrite=args.overwrite)
+        for night in np.atleast_1d(args.night):
+            process_exposures(night, verbose=args.verbose, overwrite=args.overwrite)
 
     if args.group_spectra:
         spectradir = os.path.join(outdir, 'spectra-64')
-        cmd = 'desi_group_spectra --reduxdir {outdir} --nights {night} --outdir {spectradir}'
-        cmd = cmd.format(outdir=outdir, night=str(args.night), spectradir=spectradir)
-        os.system(cmd)
+        for night in np.atleast_1d(args.night):
+            cmd = 'desi_group_spectra --reduxdir {outdir} --nights {night} --outdir {spectradir}'
+            cmd = cmd.format(outdir=outdir, night=str(night), spectradir=spectradir)
+            os.system(cmd)
 
     if args.redrock:
         specdir = os.path.join(outdir, 'spectra-64')
@@ -765,11 +637,14 @@ def main():
         os.system(cmd)
         
     if args.qaplots:
-        qa_throughput(args.night, verbose=args.verbose, overwrite=args.overwrite)
+        for night in np.atleast_1d(args.night):
+            qa_stdstars(night, verbose=args.verbose, overwrite=args.overwrite)
+            pdb.set_trace()
 
-        pdb.set_trace()
-        qa_redrock(args.night, verbose=args.verbose, overwrite=args.overwrite)
-        
+            qa_throughput(night, verbose=args.verbose, overwrite=args.overwrite)
+
+            pdb.set_trace()
+            qa_redrock(night, verbose=args.verbose, overwrite=args.overwrite)
 
 if __name__ == '__main__':
     main()
